:sparkles: Feat : 게시글등록 및 조건부랜더링
homepage의 <main>부분을 따로빼서 조건에 따라 다르게 렌더링 시켰습니다.
