+++ src/pages/HomePage.jsx
@@ -1,27 +1,37 @@
 import React from 'react'
 import { NavSearch } from '../components/navBack/NavBack'
 import { AllWrap } from '../style/commonStyle'
-import { FollowCompo } from '../components/followCompo/FollowCompo'
-import { AddBtn } from '../components/iconButton/IconButton'
 import TabMenu from '../components/tabMenu/TabMenu'
-import { Link } from 'react-router-dom'
+import { useSelector,useDispatch} from 'react-redux'
+import { useEffect } from 'react'
+import { selectAllPosts,AxiosPetInfo,getPostStatus } from '../reducers/getPetInfoSlice'
+import DefaultFeed from '../template/walkingFeed/DefaultFeed'
+import WalkingFeed from '../template/walkingFeed/WalkingFeed'
 
 export default function HomePage() {
-  const textBtn = "펫 등록하기"
-  const textDefault = "함께 놀 친구들을 찾아보세요! :)"
-  const url = '/post'
-  return (
+
+  const dispatch = useDispatch();
+  const postsStatus = useSelector(getPostStatus);
+  const URL = "https://mandarin.api.weniv.co.kr";
+  const userInfo = JSON.parse(localStorage.getItem("userinfo")).user
+  const accountname = userInfo.accountname;
+  const loginReqPath = `/product/${accountname}/?limit=30`; //내게시글
+  // const loginReqPath = `/product/?limit=100`;  // 전체게시글 100개
+  const posts = useSelector(selectAllPosts).product;
+  useEffect(()=>{
+    if(postsStatus ==='idle'){
+      dispatch(AxiosPetInfo(URL+loginReqPath))
+    }
+  },[dispatch])
+
+  return (          
     <AllWrap>
       <header>
         <NavSearch text={"산책 피드"} />
       </header>
-      <main>
-        <FollowCompo url={url} textBtn={textBtn} textDefault={textDefault} />
-        <Link to={url}>
-          <AddBtn />
-        </Link>
-        <TabMenu />
-      </main>
+      {(postsStatus==='idle'&&posts?.length===0) ?<DefaultFeed/>:<WalkingFeed/> }
+      <TabMenu />
     </AllWrap>
+              
   )
 }
